toolkit/devtools/server/tests/mochitest/test_inspector-hide.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
Components.utils.import("resource://gre/modules/devtools/Loader.jsm");

const promise = devtools.require("sdk/core/promise");
const inspector = devtools.require("devtools/server/actors/inspector");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gWalker = null;
var gClient = null;

addTest(function setup() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    let {InspectorFront} = devtools.require("devtools/server/actors/inspector");
    let inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
    }).then(runNextTest));
  });
});

addTest(function testRearrange() {
  let listFront = null;
  let listNode = gInspectee.querySelector("#longlist");

  promiseDone(gWalker.querySelector(gWalker.rootNode, "#longlist").then(front => {
    listFront = front;
  }).then(() => {
    let computed = gInspectee.defaultView.getComputedStyle(listNode);
    ok(computed.visibility, "visible", "Node should be visible to start with");
    return gWalker.hideNode(listFront);
  }).then(response => {
    let computed = gInspectee.defaultView.getComputedStyle(listNode);
    ok(computed.visibility, "hidden", "Node should be hidden");
    return gWalker.unhideNode(listFront);
  }).then(() => {
    let computed = gInspectee.defaultView.getComputedStyle(listNode);
    ok(computed.visibility, "visible", "Node should be visible again.");
  }).then(runNextTest));
});

addTest(function cleanup() {
  delete gWalker;
  delete gClient;
  runNextTest();
});

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
